TLCALIPER-107 - Change BASH script to Python program (#2)
* TLCALIPER-107 - Removed `toArray()`

Removed `toArray()` call because it's unneeded, as theorized earlier.  @esther11 thought it was slowing the process.  I didn't see any difference in speed running the program with or without that call, even with large numbers of events, but it's not good to have unnecessary code left in the program anyway.

* TLCALIPER-107 - Point to Python program

Changed symlink to point to new Python program.

* TLCALIPER-107 - New Python program

Changed BASH program to Python.

* TLCALIPER-107 - Changes for Python program

Changed documentation to use options from new Python program.